bug fix pagination of project listing page + removed some unnecessary files.
