+++ clientprojects/urls.py
@@ -2,5 +2,5 @@ from django.urls import include, path
 from . import views
 
 urlpatterns = [
-path('clientprojects', views.client_projects, name='clientprojects'),
+    path('clientprojects', views.project_view, name='clientprojects'),
 ]
