--- freelance_webpage/static/css/fancy.css
@@ -1,217 +1,0 @@
-.hire_me_popup{
-    border-radius: 5px;
-    background: #000000;
-    width: 450px;
-    padding: 30px;
-}
-.hire_me_popup_box {
-    width: 100%;
-}
-.hire_me_popup_box h4 {
-    font-size: 24px;
-    font-weight: bold;
-    margin-bottom: 20px;
-    color: #ffffff;
-}
-.hire_me_popup_box h4 svg{
-    width: 20px;
-    height: 20px;
-    fill: #ffffff;
-}
-.hire_me_popup_box form{
-    margin-top: 30px;
-}
-.hire_me_popup_box form label {
-    color: #ffffff;
-    font-size: 16px;
-    margin-bottom: 6px;
-}
-.hire_me_popup_box form input {
-    width: 100%;
-    border-radius: 5px;
-    border: none;
-    padding: 10px;
-    margin-bottom: 20px;
-}
-.hire_me_popup_box form textarea{
-    width: 100%;
-    border-radius: 5px;
-    padding: 10px;
-    margin-bottom: 20px;
-}
-.hire_me_popup_box form .form-box {
-    width: 100%;
-    display: flex;
-    justify-content: space-between;
-    margin: 15px 0;
-    flex-direction: column;
-}
-.hire_me_popup_box form select {
-    width: 100%;
-    height: 50px;
-    padding: 5px;
-    border-radius: 8px;
-}
-.hire_me_popup_box form .form-box .input_box{
-    width: 100%;
-    margin-top: 20px;
-}
-.hire_me_popup_box form .input_box input{
-    width: calc(100% - 100px);
-    height: 50px;
-    float: left;
-    border-top-left-radius: 8px;
-    border-bottom-left-radius: 8px;
-    border: 1px solid #111111;
-    border-right: none;
-    padding: 20px;
-}
-.hire_me_popup_box form .input_box button{
-    width: 100px;
-    height: 50px;
-    float: left;
-    border-top-right-radius: 8px;
-    border-bottom-right-radius: 8px;
-    border: 1px solid #111111;
-    background: none;
-}
-.hire_me_popup_box form .input_box button svg{
-    width: 20px;
-    height: 20px;
-}
-.hire_me_popup_box form  .min-budget{
-    width: 100%;
-    display: flex;
-    justify-content: space-between;
-    margin-bottom: 30px;
-    flex-direction: column;
-}
-.hire_me_popup_box form .min-budget .input_box{
-    width: 100%;
-    
-}
-.hire_me_popup_box form .min-budget .input_box input{
-    border-radius: 0;
-     border-top-right-radius: 8px;
-    border-bottom-right-radius: 8px;
-    border-right: 1px solid;
-    border-left: none;
-}
-.hire_me_popup_box form  .min-budget .input_box button{
-    border-radius: 0;
-    border-top-left-radius: 8px;
-    border-bottom-left-radius: 8px;
-}
-.hire_me_popup_box form  .min-budget select{
-    margin-top: 20px;
-}
-.hire_me_popup_box form p{
-    font-size: 16px;
-    margin-top: 20px;
-}
-.hire_me_popup_box form p a{
-    color: red;
-}
-.select-filde{
-    margin: 0;
-    padding: 0;
-    list-style-type: none;
-}
-.select-filde li{
-    display: inline-block;
-}
-.select-filde li input {
-    width: auto;
-    margin: 0;
-    margin-right: 7PX;
-}
-.select-filde li label{
-    text-transform: capitalize;
-}
-.select-filde li + li{
-    margin-left: 10px;
-}
-.product{
-    float: left;
-    width: 100%;
-    margin-bottom: 20px;
-}
-.product ul{
-    list-style-type: none;
-    margin: 0;
-    padding: 0;
-}
-.product ul li {
-    display: inline-block;
-    position: relative;
-}
-.product ul li:first-child {
-    width: 65%;
-    float: left;
-    margin-right: 5%;
-}
-.product ul li:last-child {
-    width: 30%;
-    float: left;
-    margin-top: 30px;
-}
-.product ul li:first-child  input[type="number"]{
-    padding-left: 50px;
-    margin-bottom: 0;
-}
-.product ul li span {
-    position: absolute;
-    left: 0;
-    bottom: 0;
-    width: 45px;
-    height: 45px;
-    background: #f0f0f0;
-    z-index: 9999999;
-    color: #000000;
-    text-align: center;
-    border: 1px solid #cccccc;
-    line-height: 45px;
-    border-radius: 5px;
-}
-.input-box {
-    width: 100%;
-    float: left;
-    margin-top: 20px;
-    position: relative;
-}
-.product ul li span.hr{
-    left: auto;
-    right: 0;
-}
-.input-box input{
-    margin-bottom: 0 !important;
-}
-.input-box span{
-    position: absolute;
-    bottom: 0;
-    right: 0;
-    height: 45px;
-    border-radius: 5px;
-    border: 1px solid #cccccc;
-    background: #f0f0f0;
-    color: #000000;
-    z-index: 9999;
-    text-align: center;
-    padding: 0 10px;
-    line-height: 45px;
-}
-
-.submitForm {
-    margin-top: 14px;
-    border-radius: 30px;
-    padding: 6px 14px;
-    border: 1px solid #fccf14;
-    background-color: #fccf14;
-    color: #111;
-}
-.activeTab{
-    display:block !important;
-}
-.someData{
-    display: none;
-}
