+++ freelance_webpage/templates/common/navbar.html
@@ -42,7 +42,6 @@
                                             {% else %}
                                               <li><a href="{% url 'postproject' %}">Post Project</a></li>
                                               <li><a href="{% url 'clientprojects' %}">My Project Listings</a></li>
-                                              <li><a href="#">Search Freelancers</a></li>
                                               <li><a href="{% url 'client_profile' %}">Update Profile</a></li>
                                             {% endif %}
                                         </ul>
