+++ freelance_webpage/templates/dashboards/employer-dboard.html
@@ -84,17 +84,6 @@
           </a>
         </div>
         <div class="col-md-4 mb-4">
-          <a href="#">
-            <div class="card text-center">
-              <div class="card-body">
-                <i class="fas fa-search fa-3x mb-3"></i>
-                <h5 class="card-title">Search Freelancers</h5>
-              </div>
-            </div>
-          </a>
-        </div>
-        <!-- Add two empty columns to create space for five cards in a row -->
-        <div class="col-md-4 mb-4">
             <a href="{%url 'client_profile' %}">
               <div class="card text-center">
                 <div class="card-body">
