+++ post_project/urls.py
@@ -3,5 +3,4 @@ from . import views
 
 urlpatterns = [
     path('postproject', views.post_project, name='postproject'),
-    path('clientprojects', views.project_view, name='clientprojects'),
 ]
